backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
google-generativeai==0.7.2
langgraph==0.1.19
langchain-core==0.2.10
arxiv==2.1.0
sentence-transformers==3.0.1
faiss-cpu>=1.12.0
sqlalchemy==2.0.31
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.8.2
httpx==0.27.0
